refactor : 사용하지 않는 코드 제거 후 새로운 로직으로 변경
