--- src/main/java/chess/dto/GameDto.java
@@ -1,13 +1,0 @@
-package chess.dto;
-
-public class GameDto {
-    private Long id;
-
-    public GameDto(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
