+++ src/test/java/chess/service/ChessServiceTests.java
@@ -7,27 +7,31 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import chess.dao.RoomFakeDao;
 import chess.dao.SquareFakeDao;
 import java.util.List;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-class ChessServiceV2Test {
+class ChessServiceTests {
 
-    private final ChessServiceV2 chessServiceV2 = new ChessServiceV2(new RoomFakeDao(), new SquareFakeDao());
+    private ChessService chessService = new ChessService(
+            new RoomFakeDao(),
+            new SquareFakeDao()
+    );
 
     @Test
     @DisplayName("전체 방 리스트를 조회할 수 있다.")
     void findAllRoom() {
-        chessServiceV2.insertRoom("title1", "1111");
-        chessServiceV2.insertRoom("title1", "1111");
+        chessService.insertRoom("title1", "1111");
+        chessService.insertRoom("title1", "1111");
 
-        assertThat(chessServiceV2.findAllRoom()).hasSize(2);
+        Assertions.assertThat(chessService.findAllRoom()).hasSize(2);
     }
-    
+
     @Test
     @DisplayName("한개의 방을 생성할 수 있다.")
     void insertRoom() {
-        final Long roomId = chessServiceV2.insertRoom("title1", "1111");
+        final Long roomId = chessService.insertRoom("title1", "1111");
 
         assertThat(roomId).isInstanceOf(Long.class);
     }
@@ -35,8 +39,8 @@ class ChessServiceV2Test {
     @Test
     @DisplayName("방의 일련번호를 통해 보드판과 방의 상태를 변경할 수 있다.")
     void insertBoard() {
-        final Long roomId = chessServiceV2.insertRoom("title1", "1111");
-        final Long updateRoomId = chessServiceV2.insertBoard(roomId);
+        final Long roomId = chessService.insertRoom("title1", "1111");
+        final Long updateRoomId = chessService.insertBoard(roomId);
 
         assertThat(updateRoomId).isInstanceOf(Long.class);
     }
@@ -44,28 +48,28 @@ class ChessServiceV2Test {
     @Test
     @DisplayName("방의 일련번호를 통해 전체 보드 구성요소를 조회할 수 있다.")
     void findSquareAllById() {
-        final Long roomId = chessServiceV2.insertRoom("title1", "1111");
-        chessServiceV2.insertBoard(roomId);
+        final Long roomId = chessService.insertRoom("title1", "1111");
+        chessService.insertBoard(roomId);
 
-        assertThat(chessServiceV2.findSquareAllById(roomId)).hasSize(64);
+        Assertions.assertThat(chessService.findSquareAllById(roomId)).hasSize(64);
     }
 
     @Test
     @DisplayName("방의 일련번호와 현재위치 타겟위치를 통해 보드의 구성요소를 변경할 수 있다.")
     void updateSquares() {
-        final Long roomId = chessServiceV2.insertRoom("title1", "1111");
-        chessServiceV2.insertBoard(roomId);
+        final Long roomId = chessService.insertRoom("title1", "1111");
+        chessService.insertBoard(roomId);
 
-        assertDoesNotThrow(() -> chessServiceV2.updateSquares(roomId, "a2", "a3"));
+        assertDoesNotThrow(() -> chessService.updateSquares(roomId, "a2", "a3"));
     }
 
     @Test
     @DisplayName("일련번호를 통해 방의 점수를 조회할 수 있다.")
     void findStatusById() {
-        final Long roomId = chessServiceV2.insertRoom("title", "1111");
-        chessServiceV2.insertBoard(roomId);
+        final Long roomId = chessService.insertRoom("title", "1111");
+        chessService.insertBoard(roomId);
 
-        final List<Double> status = chessServiceV2.findStatusById(roomId);
+        final List<Double> status = chessService.findStatusById(roomId);
 
         assertThat(status).hasSize(2);
     }
@@ -73,10 +77,10 @@ class ChessServiceV2Test {
     @Test
     @DisplayName("방의 상태를 종료로 변경할 수 있다.")
     void updateStateEnd() {
-        final Long roomId = chessServiceV2.insertRoom("title1", "1111");
-        chessServiceV2.insertBoard(roomId);
+        final Long roomId = chessService.insertRoom("title1", "1111");
+        chessService.insertBoard(roomId);
 
-        assertDoesNotThrow(() -> chessServiceV2.updateStateEnd(roomId));
+        org.junit.jupiter.api.Assertions.assertDoesNotThrow(() -> chessService.updateStateEnd(roomId));
     }
 
     @Nested
@@ -86,17 +90,17 @@ class ChessServiceV2Test {
         @Test
         @DisplayName("비밀번호가 일치하면 삭제된다.")
         void valid() {
-            final Long roomId = chessServiceV2.insertRoom("title1", "1111");
+            final Long roomId = chessService.insertRoom("title1", "1111");
 
-            assertDoesNotThrow(() -> chessServiceV2.deleteRoom(roomId, "1111"));
+            org.junit.jupiter.api.Assertions.assertDoesNotThrow(() -> chessService.deleteRoom(roomId, "1111"));
         }
 
         @Test
         @DisplayName("비밀번호가 일치히지 않으면 예외가 발생한다.")
         void invalidNotEqualsPassword() {
-            final Long roomId = chessServiceV2.insertRoom("title1", "1111");
+            final Long roomId = chessService.insertRoom("title1", "1111");
 
-            assertThatThrownBy(() -> chessServiceV2.deleteRoom(roomId, "1234"))
+            assertThatThrownBy(() -> chessService.deleteRoom(roomId, "1234"))
                     .isInstanceOf(IllegalArgumentException.class)
                     .hasMessage("비밀번호가 일치하지 않습니다.");
         }
@@ -104,10 +108,10 @@ class ChessServiceV2Test {
         @Test
         @DisplayName("비밀번호가 일치하지만 방의 상태가 WhiteRunning, BlackRunning일 경우 예외가 발생한다.")
         void invalidState() {
-            final Long roomId = chessServiceV2.insertRoom("title1", "1111");
-            chessServiceV2.insertBoard(roomId);
+            final Long roomId = chessService.insertRoom("title1", "1111");
+            chessService.insertBoard(roomId);
 
-            assertThatThrownBy(() -> chessServiceV2.deleteRoom(roomId, "1111"))
+            assertThatThrownBy(() -> chessService.deleteRoom(roomId, "1111"))
                     .isInstanceOf(IllegalStateException.class)
                     .hasMessage("게임이 실행중일 경우 게임을 삭제할 수 없습니다.");
         }
